task every hour

+++ task.worker.js
@@ -6,7 +6,7 @@ import lib from './lib/lib.js'
 
 self.onmessage = (event) => {
     var db = lib.sqlite(os.homedir() + "/.brook.db", { wal: true })
-    new CronJob('0 0 0 * * *', async function() {
+    async function task() {
         var l = db.query(`select * from task`).all()
         for (var i = 0; i < l.length; i++) {
             var v = l[i]
@@ -43,17 +43,30 @@ self.onmessage = (event) => {
                     await $`sshexec -s '${v.server}' -u '${v.user}' -k /tmp/sshkey --download '${v.pid_path}' --to '/tmp/_' --timeout 60`
                 }
                 var s = (await fs.readFile("/tmp/_", { encoding: 'utf8' })).trim()
-                if (v.password) {
-                    await $`sshexec -s '${v.server}' -u '${v.user}' -p '${v.password}' -c 'kill -USR1 ${s}' --timeout 60`
+                if (v.user == "root") {
+                    if (v.password) {
+                        await $`sshexec -s '${v.server}' -u '${v.user}' -p '${v.password}' -c 'kill -USR1 ${s}' --timeout 60`
+                    }
+                    if (v.sshkey) {
+                        await $`sshexec -s '${v.server}' -u '${v.user}' -k /tmp/sshkey -c 'kill -USR1 ${s}' --timeout 60`
+                    }
                 }
-                if (v.sshkey) {
-                    await $`sshexec -s '${v.server}' -u '${v.user}' -k /tmp/sshkey -c 'kill -USR1 ${s}' --timeout 60`
+                if (v.user != "root") {
+                    if (v.password) {
+                        await $`sshexec -s '${v.server}' -u '${v.user}' -p '${v.password}' --sudo -c 'kill -USR1 ${s}' --timeout 60`
+                    }
+                    if (v.sshkey) {
+                        await $`sshexec -s '${v.server}' -u '${v.user}' -k /tmp/sshkey --sudo -c 'kill -USR1 ${s}' --timeout 60`
+                    }
                 }
             } catch (e) {
                 console.log(e)
             }
         }
-    }, null, true, 'utc');
+        setTimeout(task, 60 * 60 * 1000)
+    }
+    setTimeout(task, 60 * 60 * 1000)
+
     new CronJob('0 0 0 1 * *', function() {
         db.query(`update user set traffic_now=0`).run()
     }, null, true, 'utc');
